Issue #270 - Development started
